+++ src/core/DeckBuilder/DBGridCardObject.gd
@@ -25,12 +25,12 @@ func setup(card_name) -> Card:
 	display_card = cfc.instance_card(card_name)
 	add_child(display_card)
 	display_card.resize_recursively(display_card._control, CFConst.THUMBNAIL_SCALE)
-	display_card.card_front.scale_to(CFConst.THUMBNAIL_SCALE)	
+	display_card.card_front.scale_to(CFConst.THUMBNAIL_SCALE)
 	display_card.state = Card.CardState.DECKBUILDER_GRID
 	rect_min_size = CFConst.CARD_SIZE * CFConst.THUMBNAIL_SCALE
-	rect_size = rect_min_size	
+	rect_size = rect_min_size		
 	return(display_card)
-	
+
 
 func _on_DBGridCardObject_mouse_entered() -> void:
 	preview_popup.show_preview_card(display_card.canonical_name)